Framework Target to 4.8
Add unit tests
Add util method to encrypt large string
