+++ appveyor.yml
@@ -14,15 +14,15 @@ after_build:
 
     7z a ShadowsocksR-netcore-win32.zip %APPVEYOR_BUILD_FOLDER%\shadowsocks-csharp\bin\Release\netcoreapp3.0\win-x86\publish\ShadowsocksR.exe
     
-    7z a ShadowsocksR-net472.zip %APPVEYOR_BUILD_FOLDER%\shadowsocks-csharp\bin\Release\net472\ShadowsocksR.exe
+    7z a ShadowsocksR-net48.zip %APPVEYOR_BUILD_FOLDER%\shadowsocks-csharp\bin\Release\net48\ShadowsocksR.exe
 
     7z a ShadowsocksR-netcore-win64.zip %APPVEYOR_BUILD_FOLDER%\shadowsocks-csharp\bin\Release\netcoreapp3.0\win-x64\publish\ShadowsocksR.exe
 test: off
 artifacts:
 - path: ShadowsocksR-netcore.zip
   name: ShadowsocksR-netcore.zip
-- path: ShadowsocksR-net472.zip
-  name: ShadowsocksR-net472.zip
+- path: ShadowsocksR-net48.zip
+  name: ShadowsocksR-net48.zip
 - path: ShadowsocksR-netcore-win32.zip
   name: ShadowsocksR-netcore-win32.zip
 - path: ShadowsocksR-netcore-win64.zip
